Create Player Character (#16)
* Create Temp Sprite and Implement Basic Movement

Temporary sprite is just the Godot icon, and current movement is just up,
down, left, and right using WASD or arrow keys. Speed is the default of 300 and the player is unaffected by gravity.

* Create Camera and Temp Background

A Camera2D node is now centered on the player character, and there's
temporary background text to show that the camera is in fact
following the player when they move.

* Test for Player Mortality

If DEBUG is enabled, a 5 second timer will start where the Player HP is
set to 0 at timeout. Once at 0, the Player should then be removed from
the scene tree as the Player should now be dead.

* Add Player Idle & Walk Animations

The player now has a new sprite and accompanying idle and walk
animations. Credit for the sprite assets goes to Michael Games @
https://michaelgames.itch.io/2d-action-adventure-rpg-assets?download

* Create Health and Hitbox Components

Health is now a component separated from being tied to a specific class,
and same with the hitbox component. There's a great video on YouTube by
Bitlytic titled "How You Can Easily Make Your Code Simpler in Godot 4"
that explains how and why to do this.

* Remove Temp Background

* Update gitignore

Test folder is now ignored

* Move Player and Reorganize Folders

The dungeon map has been placed as a child of the Game root node and
player has been placed on the map. The directory structure has also been
reorganized for clarity.

* Remove Empty Scene